+++ src/labthings/default_views/docs/static/favicon-16x16.png
