+++ src/labthings/default_views/docs/static/favicon-32x32.png
