+++ src/labthings/default_views/docs/static/index.html
