+++ src/labthings/default_views/docs/static/oauth2-redirect.html
