+++ src/labthings/default_views/docs/static/swagger-ui-standalone-preset.js
