+++ src/labthings/default_views/docs/templates/swagger-ui.html
