--- src/labthings/server/sockets/__init__.py
@@ -1,1 +1,0 @@
-from .gevent import Sockets, SocketSubscriber
