--- src/labthings/server/wsgi/__init__.py
@@ -1,15 +1,0 @@
-from .gevent import Server
-
-"""
-NOTE: THIS FILE IS EXCLUDED FROM OUR UNIT TESTS.
-STARTING A WSGI SERVER IN A CI IS A BIT SKETCHY AS IT
-MAKES ASSUMPTIONS ABOUT THE STATE OF THE HOST SYSTEM.
-
-FOR EXAMPLE, PORT CLASHES COULD CAUSE THE TEST TO FAIL
-DESPITE THE CODE ITSELF BEING TOTALLY FINE.
-
-THIS DOES MEAN THAT EXTRA CARE SHOULD BE TAKEN WHEN
-DEVELOPING THIS SECTION OF THE CODE.
-
-THANKS
-"""
